+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Header";
+export * from "./ItemList";
+export * from "./ComplexFrom";
+export * from "./NotificationSystem";
