+++ src/context/notification/NotificationProvider.tsx
@@ -1,0 +1,37 @@
+import { ReactNode, useState } from "react";
+import { useCallback, useMemo } from "../../@lib";
+import { INotification } from "../../types";
+import { NotificationContext } from "./NotificationContext.ts";
+
+export const NotificationProvider = ({ children }: { children: ReactNode }) => {
+  const [notifications, setNotifications] = useState<INotification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: INotification["type"]) => {
+      const newNotification: INotification = {
+        id: Date.now(),
+        message,
+        type,
+      };
+      setNotifications((prev) => [...prev, newNotification]);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  }, []);
+
+  const notificationValue = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={notificationValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
