+++ src/context/theme/ThemeProvider.tsx
@@ -1,0 +1,27 @@
+import { ReactNode, useState } from "react";
+import { useCallback, useMemo } from "../../@lib";
+import { IThemeContext } from "../../types/context";
+import { ThemeContext } from "./ThemeContext.ts";
+
+export const ThemeProvider = ({ children }: { children: ReactNode }) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  }, []);
+
+  const themeContextValue = useMemo<IThemeContext>(
+    () => ({ theme, toggleTheme }),
+    [theme, toggleTheme],
+  );
+
+  return (
+    <ThemeContext.Provider value={themeContextValue}>
+      <div
+        className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+      >
+        {children}
+      </div>
+    </ThemeContext.Provider>
+  );
+};
