+++ src/context/theme/index.ts
@@ -1,0 +1,2 @@
+export * from "./ThemeContext";
+export * from "./ThemeProvider";
