+++ src/context/user/UserContext.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from "react";
+import { IUserContext } from "../../types/context";
+
+export const UserContext = createContext<IUserContext | null>(null);
+
+export const useUser = () => {
+  const context = useContext(UserContext);
+  if (context === null) {
+    throw new Error("useUser must be used within a UserProvider");
+  }
+  return context;
+};
