+++ src/types/Item.ts
@@ -1,0 +1,6 @@
+export interface IItem {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
