+++ src/types/User.ts
@@ -1,0 +1,5 @@
+export interface IUser {
+  id: number;
+  name: string;
+  email: string;
+}
