+++ src/types/context/Notification.ts
@@ -1,0 +1,7 @@
+import { INotification } from "../Notification.ts";
+
+export interface INotificationContext {
+  notifications: INotification[] | [];
+  addNotification: (message: string, type: INotification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
