+++ src/types/context/Theme.ts
@@ -1,0 +1,4 @@
+export interface IThemeContext {
+  theme: string;
+  toggleTheme: () => void;
+}
