+++ src/types/context/User.ts
@@ -1,0 +1,7 @@
+import { IUser } from "../User.ts";
+
+export interface IUserContext {
+  user: IUser | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
