+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./User";
+export * from "./Notification";
+export * from "./Item";
